==> 1_pcr-globwb/02_create_config_files_pcr-globwb.ipynb <==
# Create config files PCR-GLOBWB
### Parameterset Location on Snellius: /gpfs/work1/0/dfguu/users/edwin/data/pcrglobwb_input_arise/develop/

from glob import glob
from pathlib import Path

from configparser import ConfigParser

# Set Paths

# Snellius paths
ROOT = Path('/gpfs/work1/0/wtrcycle/users/jaerts/camels_uk/')
MODELS = Path(f'{ROOT}/pcr-globwb/')
FORCING = Path(f'{ROOT}/forcing/')
AUXDATA = Path(f'{ROOT}/aux_data/')

# Set config

# Get available basin IDs wflow_sbm
basin_dirs = glob(f'{MODELS}/*')
basin_ids = [s.split('/')[-1] for s in basin_dirs]
basin_ids.remove('uk')
basin_ids.sort()


# Template .ini file
template = f'{AUXDATA}/PCR-GLOBWB/template_1k_setup_file.ini'

## Create UK config

basin_ids = ['uk']
for basin_id in basin_ids:
    # Load config parser
    config = ConfigParser()
    config.optionxform = str

    # Load config file
    config.read(template)

    # Set Global Options
    config['globalOptions']['outputDir'] = f'{MODELS}/{basin_id}/'
    config['globalOptions']['cloneMap'] = f'{MODELS}/{basin_id}/{basin_id}_30sec_clone.map'
    config['globalOptions']['landmask'] = f'{MODELS}/{basin_id}/{basin_id}_30sec_clone.map'
    
    # Set Period and Spin-up
    config['globalOptions']['startTime'] = "2007-01-01"
    config['globalOptions']['endTime'] = '2017-12-31'
    config['globalOptions']['maxSpinUpsInYears'] = '30'

    # Set Meteo Options 
    config['meteoOptions']['precipitationNC'] = f'{MODELS}/{basin_id}/ceh-gear_chess_camels-gb_{basin_id}_pr_clim2000-2007_2017.nc'
    config['meteoOptions']['temperatureNC']   = f'{MODELS}/{basin_id}/ceh-gear_chess_camels-gb_{basin_id}_tas_clim2000-2007_2017.nc'
    config['meteoOptions']['refETPotFileNC']  = f'{MODELS}/{basin_id}/ceh-gear_chess_camels-gb_{basin_id}_pet_clim2000-2007_2017.nc'
    
    config['meteoDownscalingOptions']['downscalePrecipitation'] = 'True'
    config['meteoDownscalingOptions']['downscaleTemperature'] = 'True' 
    config['meteoDownscalingOptions']['downscaleReferenceETPot'] = 'True'

    # Write output to basin directory
    outfile = f'{MODELS}/{basin_id}/{basin_id}.ini'
    with open(outfile, "w") as configfile:
            config.write(configfile)
