==> 1_pcr-globwb/04_pcr-globwb_run.ipynb <==
# Run PCR-GLOBWB 

### /gpfs/work1/0/dfguu/users/edwin/data/pcrglobwb_input_arise/develop/

import os
import subprocess
import pandas as pd
import xarray as xr

from glob import glob
from pathlib import Path
from pathos.threading import ThreadPool as Pool

# Set Paths

# Set Paths
ROOT = Path('/gpfs/work1/0/wtrcycle/users/jaerts/camels_uk/')
MODELS = Path(f'{ROOT}/pcr-globwb/')
SCRIPTS = Path(f'{ROOT}/code/')

# Config

# Get available basin IDs wflow_sbm
basin_dirs = glob(f'{MODELS}/*')
basin_ids = [s.split('/')[-1] for s in basin_dirs]
basin_ids.remove('uk')
basin_ids.sort()

# Set available cores
cores_available = 40

# Sort by basin size
def sort_basin_ids_by_size(basin_ids):
    sizes = []
    for basin_id in basin_ids:
        size = os.path.getsize(f'{MODELS}/{basin_id}/{basin_id}_30sec_nc_clone.nc')
        sizes.append(size)

    df = pd.DataFrame()
    df['basin_id'] = basin_ids
    df['size'] = sizes
    df = df.sort_values('size')

    basin_ids = df.basin_id.to_list()
    
    return basin_ids

basin_ids_sorted = sort_basin_ids_by_size(basin_ids)

# Define runner and parallel functions

def pcrglobwb_runner(basin_id):
    print(f'Starting: {basin_id}')

    # Set config_file
    config_file = f'{MODELS}/{basin_id}/{basin_id}.ini'
    # Set run script
    run_script = f'{SCRIPTS}/pcr-globwb/PCR-GLOBWB_model/model/deterministic_runner.py'
    print(f'python {run_script} {config_file}')

    subprocess.call(
                    f'python {run_script} {config_file}',
                    stdout=subprocess.DEVNULL,
                    stderr=subprocess.STDOUT,
                    shell=True
                   )

    print(f'Finished: {basin_id}')

def parallel_run(basin_ids, threads=cores_available):
    
    # Set number of threads (cores) used for parallel run and map threads
    if threads is None:
        pool = Pool()
    else:
        pool = Pool(nodes=threads)
        
    # Run parallel models
    pool.map(pcrglobwb_runner, basin_ids)
    return

# Run parallel

parallel_run(basin_ids)

# Run UK

basin_id = 'uk'
# Set config_file
config_file = f'{MODELS}/{basin_id}/{basin_id}_spinup.ini'
# Set run script
run_script = f'{SCRIPTS}/pcr-globwb/PCR-GLOBWB_model/model/deterministic_runner.py'
print(f'python {run_script} {config_file}')

subprocess.call(
                f'python {run_script} {config_file}',
                # stdout=subprocess.DEVNULL,
                # stderr=subprocess.STDOUT,
                shell=True
               )
